+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v1.1.1
+- Fix error if multiple root selectors are defined
+- Fix incorrect output with shorthand properties
+
 ## v1.1.0
 - Use [`node.error`](https://github.com/postcss/postcss/blob/master/docs/guidelines/plugin.md#31-use-nodeerror-on-css-relevant-errors) for errors.
 - Correct readme for installing.